Added New Event UI & filters to sort events Also fixed couple places and added new methods to extensions
