+++ Pet Reminder/Views/Events/List/TodaysEventsView.swift
@@ -1,5 +1,5 @@
 //
-//  FutureEventsView.swift
+//  TodaysEventsView.swift
 //  Pet Reminder
 //
 //  Created by Ege Sucu on 27.04.2023.
@@ -9,12 +9,23 @@
 import SwiftUI
 import EventKit
 
-struct FutureEventsView: View {
+struct TodaysEventsView: View {
 
     var eventVM: EventManager
+    
+    @Binding var filteredCalendar: EKCalendar?
 
     var filteredEvents: [EKEvent] {
-        eventVM.events.filter({ Calendar.current.isDateLater(date: $0.startDate) })
+        withAnimation {
+            eventVM.events.filter {
+                if let filteredCalendar {
+                    return Calendar.current.isDateInToday($0.startDate) &&
+                    $0.calendar == filteredCalendar
+                } else {
+                    return Calendar.current.isDateInToday($0.startDate)
+                }
+            }
+        }
     }
 
     var body: some View {
@@ -29,11 +40,11 @@ struct FutureEventsView: View {
                 }
             }
         } header: {
-            Text("upcoming_title")
+            Text("today_title")
         }
     }
 }
 
 #Preview {
-    FutureEventsView(eventVM: .init(isDemo: true))
+    TodaysEventsView(eventVM: .init(isDemo: true), filteredCalendar: .constant(nil))
 }
